Check conditional metadata rules test with env variable

+++ cloudinary-test-common/src/main/java/com/cloudinary/test/AbstractStructuredMetadataTest.java
@@ -6,6 +6,7 @@ import com.cloudinary.api.ApiResponse;
 import com.cloudinary.api.exceptions.BadRequest;
 import com.cloudinary.metadata.*;
 
+import com.cloudinary.test.helpers.Feature;
 import com.cloudinary.utils.ObjectUtils;
 import org.hamcrest.Matchers;
 import org.junit.*;
@@ -311,52 +312,56 @@ public abstract class AbstractStructuredMetadataTest extends MockableTest {
     }
 
     @Test
-    @Ignore("Skip test, need to enable conditional metadata rules")
     public void testListMetadataRules() throws Exception {
-        ApiResponse result = cloudinary.api().listMetadataRules(null);
-        assertNotNull(result);
+        if (MockableTest.shouldTestFeature(Feature.CONDITIONAL_METADATA_RULES)) {
+            ApiResponse result = cloudinary.api().listMetadataRules(null);
+            assertNotNull(result);
+        }
     }
 
     @Test
-    @Ignore("Skip test, need to enable conditional metadata rules")
     public void testAddMetadataRule() throws Exception {
-        SetMetadataField field = createSetField("test123");
-        ApiResponse response = addFieldToAccount(field);
-        assertNotNull(response);
+        if (MockableTest.shouldTestFeature(Feature.CONDITIONAL_METADATA_RULES)) {
+            SetMetadataField field = createSetField("test123");
+            ApiResponse response = addFieldToAccount(field);
+            assertNotNull(response);
 
-        String externalId = (String) response.get("external_id");
-        MetadataRule rule = new MetadataRule(externalId, "category-employee", new MetadataRuleCondition("category", false, null, "employee"), new MetadataRuleResult(true, "all", null, null));
-        ApiResponse result = cloudinary.api().addMetadataRule(rule, ObjectUtils.asMap());
-        assertNotNull(result);
+            String externalId = (String) response.get("external_id");
+            MetadataRule rule = new MetadataRule(externalId, "category-employee", new MetadataRuleCondition("category", false, null, "employee"), new MetadataRuleResult(true, "all", null, null));
+            ApiResponse result = cloudinary.api().addMetadataRule(rule, ObjectUtils.asMap());
+            assertNotNull(result);
 
-        String name = (String) result.get("name");
-        assertEquals(name, "category-employee");
+            String name = (String) result.get("name");
+            assertEquals(name, "category-employee");
+        }
 
     }
 
     @Test
-    @Ignore("Skip test, need to enable conditional metadata rules")
     public void testUpdateMetadataRule() throws Exception {
-        ApiResponse response = cloudinary.api().listMetadataRules(null);
-        List metadataRules = (List)response.get("metadata_rules");
-        assertNotNull(metadataRules);
-        String externalId = (String) ((Map) metadataRules.get(0)).get("external_id");
+        if (MockableTest.shouldTestFeature(Feature.CONDITIONAL_METADATA_RULES)) {
+            ApiResponse response = cloudinary.api().listMetadataRules(null);
+            List metadataRules = (List) response.get("metadata_rules");
+            assertNotNull(metadataRules);
+            String externalId = (String) ((Map) metadataRules.get(0)).get("external_id");
 
-        MetadataRule rule = new MetadataRule(null, "test_name", null, null);
-        ApiResponse result = cloudinary.api().updateMetadataRule(externalId, rule, ObjectUtils.asMap());
-        assertNotNull(result);
+            MetadataRule rule = new MetadataRule(null, "test_name", null, null);
+            ApiResponse result = cloudinary.api().updateMetadataRule(externalId, rule, ObjectUtils.asMap());
+            assertNotNull(result);
+        }
     }
 
     @Test
-    @Ignore("Skip test, need to enable conditional metadata rules")
     public void testDeleteMetadataRule() throws Exception {
-        ApiResponse response = cloudinary.api().listMetadataRules(null);
-        List metadataRules = (List)response.get("metadata_rules");
-        assertNotNull(metadataRules);
-        String externalId = (String) ((Map) metadataRules.get(0)).get("external_id");
+        if (MockableTest.shouldTestFeature(Feature.CONDITIONAL_METADATA_RULES)) {
+            ApiResponse response = cloudinary.api().listMetadataRules(null);
+            List metadataRules = (List) response.get("metadata_rules");
+            assertNotNull(metadataRules);
+            String externalId = (String) ((Map) metadataRules.get(0)).get("external_id");
 
-        ApiResponse result = cloudinary.api().deleteMetadataRule(externalId, ObjectUtils.emptyMap());
-        assertNotNull(result);
+            ApiResponse result = cloudinary.api().deleteMetadataRule(externalId, ObjectUtils.emptyMap());
+            assertNotNull(result);
+        }
     }
 
     // Metadata test helpers
